Fixing client initialisation only when region is marked as object store capable
