+++ fiware-region-sanity-tests/tests/regions/test_spain.py
@@ -24,9 +24,8 @@
 __author__ = 'jfernandez'
 
 
-from tests import fiware_region_without_networks_tests, fiware_region_object_storage_tests
+from tests import fiware_region_without_networks_with_storage_tests
 
 
-class TestSuite(fiware_region_without_networks_tests.FiwareRegionWithoutNetworkTest,
-                fiware_region_object_storage_tests.FiwareRegionsObjectStorageTests):
+class TestSuite(fiware_region_without_networks_with_storage_tests.FiwareRegionWithoutNetworkTest):
     region_name = "Spain"
